+++ django_pain/tests/test_models.py
@@ -1,9 +1,19 @@
 """Test models."""
 from django.core.exceptions import ValidationError
-from django.test import SimpleTestCase
+from django.db.models import BLANK_CHOICE_DASH
+from django.test import SimpleTestCase, override_settings
 from djmoney.money import Money
 
 from django_pain.models import BankAccount, BankPayment
+
+
+class TestBankAccount(SimpleTestCase):
+    """Test BankAccount model."""
+
+    def test_str(self):
+        """Test string representation."""
+        account = BankAccount(account_number='123', currency='USD')
+        self.assertEqual(str(account), '123')
 
 
 class TestBankPayment(SimpleTestCase):
@@ -22,3 +32,9 @@ class TestBankPayment(SimpleTestCase):
         with self.assertRaises(ValidationError, msg='Bank payment PAYMENT is in different currency (CZK) '
                                                     'than bank account 123 (USD).'):
             payment.clean()
+
+    @override_settings(PAIN_PROCESSORS=['django_pain.tests.utils.DummyPaymentProcessor'])
+    def test_objective_choices(self):
+        self.assertEqual(BankPayment.objective_choices(), BLANK_CHOICE_DASH + [
+            ('django_pain.tests.utils.DummyPaymentProcessor', 'Dummy objective'),
+        ])