+++ django_pain/tests/utils.py
@@ -4,7 +4,31 @@ from typing import Any
 
 from djmoney.money import Money
 
-from django_pain.models import BankPayment
+from django_pain.models import BankAccount, BankPayment
+from django_pain.processors import AbstractPaymentProcessor
+
+
+class DummyPaymentProcessor(AbstractPaymentProcessor):
+    """Dummy payment processor."""
+
+    default_objective = 'Dummy objective'
+
+    def process_payments(self, payments):
+        """Dummy function."""
+
+    def assign_payment(self, payment, client_id):
+        """Dummy function."""
+
+
+def get_account(**kwargs: Any) -> BankAccount:
+    """Create bank account object."""
+    default = {
+        'account_number': '123456/0300',
+        'account_name': 'Account',
+        'currency': 'CZK',
+    }
+    default.update(kwargs)
+    return BankAccount(**default)
 
 
 def get_payment(**kwargs: Any) -> BankPayment:
@@ -14,6 +38,7 @@ def get_payment(**kwargs: Any) -> BankPayment:
         'account': None,
         'transaction_date': date(2018, 5, 9),
         'counter_account_number': '098765/4321',
+        'counter_account_name': 'Another account',
         'amount': Money('42.00', 'CZK'),
     }
     default.update(kwargs)