+++ django_pain/urls.py
@@ -1,9 +1,8 @@
 """django_pain url dispatcher."""
+from django.contrib import admin
 from django.urls import path
-
-from .views import PaymentListView
 
 app_name = 'pain'
 urlpatterns = [
-    path('payments/', PaymentListView.as_view(), name='payments-list'),
+    path('admin/', admin.site.urls),
 ]